+++ src/lib/components/modal/list/EditListName.tsx
@@ -80,7 +80,8 @@ export function EditListNameModal({
         handleSave();
         if (reroute)
           navigate({
-            pathname: `/contract-list/${formatSlugName(listName)}`,
+            pathname: "/contract-list/[slug]",
+            query: { slug: formatSlugName(listName) },
             replace: true,
           });
       }}