+++ src/lib/services/contractService.ts
@@ -81,7 +81,6 @@ export const useInstantiatedListByUserQuery = (
   walletAddr: Option<HumanAddr>
 ): UseQueryResult<ContractLocalInfo[]> => {
   const { indexerGraphClient } = useCelatoneApp();
-
   const queryFn = useCallback(async () => {
     if (!walletAddr)
       throw new Error(
@@ -104,10 +103,7 @@ export const useInstantiatedListByUserQuery = (
   return useQuery(
     ["instantiated_list_by_user", walletAddr, indexerGraphClient],
     queryFn,
-    {
-      keepPreviousData: true,
-      enabled: !!walletAddr,
-    }
+    { enabled: !!walletAddr }
   );
 };
 
@@ -165,6 +161,7 @@ export const useInstantiateDetailByContractQuery = (
     queryFn,
     {
       keepPreviousData: true,
+      enabled: !!contractAddress,
     }
   );
 };